implement all basic functions except add contact
